createStartmodelKim can now also export text file
There are new options to executable createStartmodelKim: -otxt , -ovtk controlling the
(independent) export to text-based model file and vtk files (option -bin was renamed -binvtk).
This change of functionality is documented in the ASKI user manual.
In the ASKI user manual, the description of the text-based file format for
exported model files was moved to a dedicated section in chapter "Files".
This section is also referred to in the ASKI developer's manual when talking
about importing an inverted model into the forward code.